moved token logic to welcome component
