+++ src/App.js
@@ -5,30 +5,15 @@ import { Route } from 'react-router-dom';
 import AddItem from './components/AddItem';
 import List from './components/List';
 import Welcome from './components/Welcome';
-import { Navigation } from './components/Navigation'
-
-import GetToken from './lib/token.js';
+import { Navigation } from './components/Navigation';
 
 function App() {
   const [token, setToken] = useState('');
   let history = useHistory();
   const push = history.push;
 
-  const handleClick = () => {
-    const newToken = GetToken();
-    saveToken(newToken);
-    setToken(newToken);
-    push('/list');
-  };
-
   const checkForToken = () => {
     return localStorage.getItem('userToken') ? true : false;
-  };
-
-  // token is saved locally, but not connected to firebase
-  const saveToken = token => {
-    localStorage.setItem('userToken', token);
-    console.log('token saved!');
   };
 
   useEffect(() => {
@@ -43,7 +28,7 @@ function App() {
       </header>
       <main>
         <Route exact path="/">
-          <Welcome handleClick={handleClick} />
+          <Welcome />
         </Route>
         <Route path="/list" component={List} />
         <Route path="/addItem" component={AddItem} />
