+++ src/components/Welcome.js
@@ -1,14 +1,36 @@
-import React from 'react';
+import React, { useState, useHistory } from 'react';
 
-const Welcome = ({ handleClick }) => {
+import GetToken from '../lib/token.js';
+
+// need to create a form handling the tokens
+// https://rangle.io/blog/simplifying-controlled-inputs-with-hooks/
+
+const Welcome = () => {
+  const [token, setToken] = useState('');
+  //let history = useHistory();
+  //const push = history.push;
+
+  // token is saved locally, but not connected to firebase
+  const saveToken = token => {
+    localStorage.setItem('userToken', token);
+    console.log('token saved!');
+  };
+
+  const createNewList = () => {
+    const newToken = GetToken();
+    saveToken(newToken);
+    setToken(newToken);
+    //push('/list');
+  };
+
   return (
     <div>
       <h2>Welcome to ShoppingList</h2>
-      <button onClick={handleClick}>New List</button>
+      <button onClick={createNewList}>New List</button>
       <label>
         Join List
-        <input type="text" placeholder="XXXX-XXXX-XXXX" />
-        <button>Join</button>
+        <input type="text" placeholder="litton pawn tilth" />
+        <button onClick={e => saveToken(e.target.value)}>Join</button>
       </label>
     </div>
   );
